changed get_samples to work with floats -> reduced number of casts

+++ src/utils/wave_parser.cpp
@@ -82,7 +82,7 @@ void taylortrack::utils::WaveParser::parse_file() {
 }
 
 
-std::string taylortrack::utils::WaveParser::get_samples(unsigned int sample_num) {
+std::string taylortrack::utils::WaveParser::get_samples(unsigned long sample_num) {
     unsigned long transfered_sample_size = ((this->block_align * sample_num) <= this->data_size ? this->block_align * sample_num : this->data_size);
     char samples[transfered_sample_size];
     for (unsigned int i = 0; i < transfered_sample_size; ++i) {
